Maintain the order of the ansible_play_hosts
The order of the ansible_play_hosts is not fixed so it everytime
it changes the order if the hosts in rabbitmq.config which triggers
rabbitmq service restart. If there are no changes we should avoid
triggering the restart.

This PR fixes the order of the ansible_play_hosts to be sorted in
reverse order so that it will be maintained for every run.

NOTE: These is no motivation behind using `reverse` order. We can use
any type of order to maintain the hosts order for every run.